добавлены комментарии в формате reST
